+++ Crucible/Code/QSBadgeImage.m
@@ -48,7 +48,8 @@
 
 
 
-- (void)drawInRect:(NSRect)rect fromRect:(NSRect)fromRect operation:(NSCompositingOperation)op fraction:(float)delta{
+- (void)drawInRect:(NSRect)rect fromRect:(NSRect)fromRect operation:(NSCompositingOperation)op fraction:(CGFloat)delta
+{
 	[super drawInRect:rect fromRect:rectFromSize([self size]) operation:op fraction:delta];
 	
 	NSString *numString=[NSString stringWithFormat:@"%d",count];
