+++ Crucible/Code/QSShading.m
@@ -24,7 +24,7 @@ struct QSGradientPoint{
 */
 
 
-void QSColorFade(void *info, const float *in, float *out){
+void QSColorFade(void *info, const CGFloat *in, CGFloat *out){
 	float v=*in;
 	float *colors=info;
 	int i;
@@ -32,13 +32,15 @@ void QSColorFade(void *info, const float *in, float *out){
 		*out++ = colors[i]*(1-v)+colors[i+4]*(v);
 	
 }
-void QSFillRectWithGradientFromEdge(NSRect rect,NSColor *start,NSColor *end,NSRectEdge startEdge){
+
+void QSFillRectWithGradientFromEdge(NSRect rect, NSColor *start, NSColor *end, NSRectEdge startEdge)
+{
 	CGColorSpaceRef colorspace = CGColorSpaceCreateDeviceRGB();
 	start=[start colorUsingColorSpaceName:NSDeviceRGBColorSpace];
 	end=[end colorUsingColorSpaceName:NSDeviceRGBColorSpace];
 	size_t components;
-	static const float domain[2] = { 0, 1 };
-	static const float range[10] = { 0, 1, 0, 1, 0, 1, 0, 1, 0, 1 };
+	static const CGFloat domain[2] = { 0, 1 };
+	static const CGFloat range[10] = { 0, 1, 0, 1, 0, 1, 0, 1, 0, 1 };
 	static const CGFunctionCallbacks callbacks = { 0, &QSColorFade, NULL };
